+++ test/test_FileView.py
@@ -14,8 +14,9 @@ class FileView_test(unittest.TestCase):
 
     def setUp(self):
         if glb.mainWindow() is None:
-            shutil.rmtree(home() + "/.lys")
-            glb.createMainWindow(show=False)
+            if os.path.exists(home() + "/.lys"):
+                shutil.rmtree(home() + "/.lys")
+            glb.createMainWindow(show=False, restore=True)
         os.makedirs(self.path, exist_ok=True)
 
     def tearDown(self):
